+++ src/main/java/io/github/astrapi69/swing/menu/model/MenuInfo.java
@@ -24,6 +24,7 @@
  */
 package io.github.astrapi69.swing.menu.model;
 
+import io.github.astrapi69.swing.menu.builder.JMenuItemInfo;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
@@ -34,6 +35,8 @@ import lombok.ToString;
 import lombok.experimental.FieldDefaults;
 import lombok.experimental.SuperBuilder;
 import io.github.astrapi69.swing.menu.KeyStrokeInfo;
+
+import java.awt.event.ActionListener;
 
 /**
  * The class {@link MenuInfo} is intended for store the information of a menu and on need to restore
@@ -56,4 +59,12 @@ public class MenuInfo
 	KeyStrokeInfo keyStrokeInfo;
 	String name;
 	String actionId;
+	boolean item;
+
+	public JMenuItemInfo toJMenuItemInfo(ActionListener actionListener)
+	{
+		return JMenuItemInfo.builder().actionListener(actionListener).text(this.text)
+			.mnemonic(this.mnemonic).keyStroke(this.keyStrokeInfo.toKeyStroke()).name(this.name)
+			.build();
+	}
 }